# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is boilerplate code for golang project",
        "title": "Go-Boilerplate",
        "contact": {
            "name": "Rachmat Adi Prakoso",
            "email": "rachmat.adi.p@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "paths": {
...
        "/checkouts/{uuid}": {
            "get": {
                "description": "API for get checkout detail",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Checkouts"
                ],
                "summary": "Get Checkout Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "checkout uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.DefaultResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/checkout.CheckoutResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "checkout.AppliedPromotionResponse": {
            "type": "object",
            "properties": {
                "discount_cents": {
                    "type": "integer"
                },
                "promotion_code": {
                    "type": "string"
                },
                "promotion_name": {
                    "type": "string"
                },
                "promotion_type": {
                    "type": "string"
                }
            }
        },
...
        "checkout.CheckoutItemResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "total_price_cents": {
                    "type": "integer"
                },
                "unit_price_cents": {
                    "type": "integer"
                }
            }
        },
        "checkout.CheckoutResponse": {
            "type": "object",
            "properties": {
                "applied_promotions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/checkout.AppliedPromotionResponse"
                    }
                },
                "checkout_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount_cents": {
                    "type": "integer"
                },
                "formatted_total": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/checkout.CheckoutItemResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "subtotal_cents": {
                    "type": "integer"
                },
                "total_cents": {
                    "type": "integer"
                }
            }
        },
...
        "response.DefaultResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}